+++ Interface/Sources/Runtime/TableView/TableViewBase.swift
@@ -94,6 +94,8 @@ open class TableViewBase<MODEL, ACTION>: ConfigurableHyperViewBase, ComposableHy
         setupConstraints()
 
         afterInit()
+
+        state.owner = self
     }
 
     private func afterInit() {